requirements.txt
# TrendRadar + OpenClaw 集成依赖

# Python 版本要求：>= 3.8

# TrendRadar 依赖
requests>=2.32.5,<3.0.0
pytz>=2025.2,<2026.0
PyYAML>=6.0.3,<7.0.0
fastmcp>=2.12.0,<2.14.0
websockets>=13.0,<14.0
boto3>=1.35.0,<2.0.0
feedparser>=6.0.0,<7.0.0
litellm>=1.57.0,<2.0.0
tenacity==8.5.0

# 额外依赖
python-dateutil>=2.8.0
